% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.addThePlatformInFile}
\alias{model.addThePlatformInFile}
\title{Title}
\usage{
model.addThePlatformInFile(path, userGo, platformNameGo, mdp, adressPlatform,
  numPlatform)
}
\arguments{
\item{path}{}

\item{userGo}{}

\item{platformNameGo}{}

\item{mdp}{}

\item{adressPlatform}{}

\item{numPlatform}{}
}
\description{
Title
}

