% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.currentVersion}
\alias{model.currentVersion}
\title{Title}
\usage{
model.currentVersion(path, jobs, thePlatform)
}
\arguments{
\item{path}{}

\item{jobs}{}

\item{thePlatform}{}
}
\description{
Title
}

