% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.JobRPlatform}
\alias{model.JobRPlatform}
\title{Recover jobs R in select platform and Add informations in file}
\usage{
model.JobRPlatform(path)
}
\arguments{
\item{path}{Path to local persistent Saagie folder}
}
\description{
Recover jobs R in select platform and Add informations in file
}

