% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.newVersion}
\alias{model.newVersion}
\title{Title}
\usage{
model.newVersion(input, thePlatform, idJob, fileName)
}
\arguments{
\item{input}{}

\item{thePlatform}{}

\item{idJob}{}

\item{fileName}{}
}
\description{
Title
}

