% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.postUpgrade}
\alias{model.postUpgrade}
\title{Title}
\usage{
model.postUpgrade(thePlatform, idJob, fileName)
}
\arguments{
\item{thePlatform}{}

\item{idJob}{}

\item{fileName}{}
}
\description{
Title
}

