% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.readNamePlatform}
\alias{model.readNamePlatform}
\title{Read a file containing the different platform where the user access}
\usage{
model.readNamePlatform(path)
}
\arguments{
\item{path}{Path to local persistent Saagie folder}
}
\value{
res is a file containing the different platform where the user access
}
\description{
Read a file containing the different platform where the user access
}

