% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.readTableJob}
\alias{model.readTableJob}
\title{Read a file containing job R or Create File}
\usage{
model.readTableJob(path)
}
\arguments{
\item{path}{Path to local persistent Saagie folder}
}
\value{
res is a file containing the list of job
}
\description{
Read a file containing job R or Create File
}

