% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.readTablePlatform}
\alias{model.readTablePlatform}
\title{Read a file where there are the different Platform name or Create File}
\usage{
model.readTablePlatform(path)
}
\arguments{
\item{path}{path where is a file "platform.csv"}
}
\value{
res is a file containing the list of platform
}
\description{
Read a file where there are the different Platform name or Create File
}

