% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.readThePlatform}
\alias{model.readThePlatform}
\title{Read a file containing the Platform where the job is upload}
\usage{
model.readThePlatform(path)
}
\arguments{
\item{path}{Path to local persistent Saagie folder}
}
\value{
res is a file containing the Platform where the job is upload
}
\description{
Read a file containing the Platform where the job is upload
}

