% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.removeLinkedNoR}
\alias{model.removeLinkedNoR}
\title{Remove no script R and displays the name of script R}
\usage{
model.removeLinkedNoR(jobs)
}
\arguments{
\item{jobs}{We don't know yet.}
}
\value{
We don't know yet.
}
\description{
Remove no script R and displays the name of script R
}

