% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.runJob}
\alias{model.runJob}
\title{Run a Job}
\usage{
model.runJob(thePlatform, reponseAdd)
}
\arguments{
\item{thePlatform}{don't know yet}

\item{reponseAdd}{don't know yet}
}
\value{
don't know yet
}
\description{
Run a Job
}

