% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.upgradeJob}
\alias{model.upgradeJob}
\title{Title}
\usage{
model.upgradeJob(path_to_persistent_saagie_files, input, idJob, pathNameFile)
}
\arguments{
\item{path_to_persistent_saagie_files}{}

\item{input}{}

\item{idJob}{}

\item{pathNameFile}{}
}
\description{
Title
}

