% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Saagie}
\alias{Saagie}
\title{Saagie Addin Main Function}
\usage{
Saagie(data, xvar, yvar)
}
\arguments{
\item{data}{We don't know yet.}

\item{xvar}{We don't know yet.}

\item{yvar}{We don't know yet.}
}
\value{
We don't know yet.
}
\description{
Saagie Addin Main Function
}

