% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validator.fileNamesPlatformConform}
\alias{validator.fileNamesPlatformConform}
\title{Title}
\usage{
validator.fileNamesPlatformConform(res, file)
}
\arguments{
\item{res}{}

\item{file}{}
}
\description{
Title
}

