% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validator.infoJob}
\alias{validator.infoJob}
\title{Check if all fields (containing job information) are OK}
\usage{
validator.infoJob(input)
}
\arguments{
\item{input}{We don't know yet.}
}
\value{
We don't know yet.
}
\description{
Check if all fields (containing job information) are OK
}

