% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view.barSelectPlatform}
\alias{view.barSelectPlatform}
\title{Details of bar page "Select a platform"}
\usage{
view.barSelectPlatform()
}
\value{
We don't know yet.
}
\description{
Details of bar page "Select a platform"
}

