% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.JobRPlatform}
\alias{model.JobRPlatform}
\title{Add the Job R in file (Depends on the platform)}
\usage{
model.JobRPlatform(path)
}
\arguments{
\item{path}{Path to local persistent Saagie folder}
}
\value{
We don't know yet.
}
\description{
Add the Job R in file (Depends on the platform)
}

