% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.readNumJob}
\alias{model.readNumJob}
\title{Read a file containing the num where upgrade the job}
\usage{
model.readNumJob(path)
}
\arguments{
\item{path}{path where is a file "row.csv"}
}
\value{
res is a file containing the num where upgrade the job
}
\description{
Read a file containing the num where upgrade the job
}

