% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.updateTablePlatform}
\alias{model.updateTablePlatform}
\title{Update "Table Platform" when the new information about the platform is writing}
\usage{
model.updateTablePlatform(path, input)
}
\arguments{
\item{path}{Path to local persistent Saagie folder}

\item{input}{We don't know yet.}
}
\description{
Update "Table Platform" when the new information about the platform is writing
}

